scum-airdrop-bot/requirements.txt
streamlit==1.35.0
python-dotenv>=1.0.0
mcrcon==0.7
discord.py==2.3.2
